feature/ split one js file into some and make ability to bundle them into one
