+++ index.html
@@ -21,6 +21,6 @@
             <small><a href="https://newsapi.org/">Powered by News API</a></small>
         </div>
     </footer>
-    <script src="./dist/main.js"></script>
+    <script src="./dist/index.js"></script>
 </body>
 </html>
